Fix: recursion json

+++ blog-application-spring/src/main/java/com/blog/model/Post.java
@@ -11,13 +11,15 @@ import javax.persistence.Id;
 import javax.persistence.OneToMany;
 import javax.validation.constraints.NotNull;
 
+import com.fasterxml.jackson.annotation.JsonIgnore;
+
 import lombok.Data;
 @Data
 @Entity
 public class Post {
 	@Id
 	@GeneratedValue(strategy = GenerationType.AUTO)
-	private Long id;
+	private Integer id;
 	@NotNull
 	private String title;
 	@NotNull
@@ -26,5 +28,6 @@ public class Post {
 	private String content;
 
 	@OneToMany(mappedBy = "post", cascade = CascadeType.ALL)
+	
 	private Set<Comment> comment = new HashSet<>();
 }
